Changed behaviour of linked list method clone().
clone() no longer destroys the dest list, but overwrites it with a new list.
dest list no longer needs to be initialised to call this function.
But if dest list is not empty when cloned to, the original list is lost and there would be a memory leak its original list nodes.